Updated FetchData

+++ app/src/main/java/com/example/thebookspot/Book.java
@@ -5,8 +5,8 @@ private final String bookTitle,author;
 private final int bookImage;
 private final String bookCost;
 private final String webLink;
-private final double avgRating;
-public Book(String bookTitle, int bookImage, String author, String bookCost,String webLink,double avgRating){
+private final float avgRating;
+public Book(String bookTitle, int bookImage, String author, String bookCost,String webLink,float avgRating){
     this.author=author;
     this.bookCost=bookCost;
     this.bookTitle=bookTitle;
@@ -31,7 +31,7 @@ this.avgRating=avgRating;
         return author;
     }
 
-    public double getAvgRating() {
+    public float getAvgRating() {
         return avgRating;
     }
